[EJBTHREE-1026] More locking tweaks; initial replication fixes
git-svn-id: https://svn.jboss.org/repos/jbossas/projects/ejb3/branches/cluster-dev@70922 84be2c1e-ba19-0410-b317-a758671a6fc1

+++ ejb3-cache/src/test/java/org/jboss/ejb3/test/cache/mock/MockCacheItem.java
@@ -25,8 +25,9 @@ package org.jboss.ejb3.test.cache.mock;
 import org.jboss.ejb3.cache.api.CacheItem;
 
 /**
+ * Mock implementation of a CacheItem.
+ * 
  * @author Brian Stansberry
- *
  */
 public class MockCacheItem 
    extends MockIdentifiable
@@ -37,7 +38,7 @@ public class MockCacheItem
    
    private boolean modified;
 
-   public MockCacheItem(long id)
+   public MockCacheItem(int id)
    {
       super(id);
    }
